# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payments Gate",
        "description": "API for handling payment transactions",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create-payment-link": {
            "post": {
                "tags": [
                    "Payment Links"
                ],
                "summary": "Create payment link",
                "operationId": "100bd203dadd22b6214b62288bf9880c",
                "parameters": [
                    {
                        "name": "X-API-KEY",
                        "in": "header",
                        "description": "Merchant API key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "api-key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "currency",
                                    "notificationUrl",
                                    "returnUrl",
                                    "expiresAt"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 199.99
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "PLN"
                                    },
                                    "notificationUrl": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://test.payment-gate.pl/callback"
                                    },
                                    "returnUrl": {
                                        "type": "string",
                                        "format": "uri",
                                        "example": "https://test.payment-gate.pl/return-url"
                                    },
                                    "expiresAt": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-03-01 12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment link created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "paymentLink": {
                                            "type": "string",
                                            "example": "http://payments-gate.pl/payment/17a4186a-9ba1-493d-aae6-d9c2918e7085"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Payment link generation failed"
                    }
                }
            }
        },
        "/api/payment/{payment_link_id}": {
            "get": {
                "tags": [
                    "Payment Links"
                ],
                "summary": "Get payment link details",
                "operationId": "a60fd6aebf6e8a350c1881f01c3f8572",
                "parameters": [
                    {
                        "name": "paymentLinkId",
                        "in": "path",
                        "description": "Payment link UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "c6e2c816-3f5e-417b-9e91-a794223aa903"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment link details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "payment": {
                                            "properties": {
                                                "paymentLinkId": {
                                                    "type": "string",
                                                    "example": "d0947206-f142-44d1-af78-7b6b865f7394"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 199.99
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "PLN"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "transaction": {
                                            "properties": {
                                                "status": {
                                                    "type": "string",
                                                    "example": "SUCCESS"
                                                },
                                                "amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 199.99
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "PLN"
                                                },
                                                "paymentMethod": {
                                                    "type": "string",
                                                    "example": "TPAY"
                                                },
                                                "returnUrl": {
                                                    "type": "string",
                                                    "format": "uri",
                                                    "example": "https://test.payment-gate.pl/return-url"
                                                }
                                            },
                                            "type": "object",
                                            "nullable": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Payment link not found"
                    },
                    "410": {
                        "description": "Payment link expired"
                    },
                    "500": {
                        "description": "Invalid payment link"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Wprowadź swój klucz API",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "api_key": []
        }
    ]
}